feat: finalizado versao estavel do middleware
