+++ .gitignore
@@ -1,7 +1,6 @@
 node_modules/
 
 dist/
-lib/
 
 .env
