+++ tsconfig.json
@@ -4,17 +4,23 @@
     "declaration": true, /* Generate corresponding '.d.ts' file. */
     "experimentalDecorators": true, /* Enable experimental support for legacy experimental decorators. */
     "emitDecoratorMetadata": true, /* Emit design-type metadata for decorated declarations in source files. */
-    "module": "ES2020", /* Specify what module code is generated. */
+    "module": "NodeNext", /* Specify what module code is generated. */
     "rootDir": "./src", /* Specify the root folder within your source files. */
-    "moduleResolution": "node", /* Specify how TypeScript looks up a file from a given module specifier. */
+    "moduleResolution": "nodenext", /* Specify how TypeScript looks up a file from a given module specifier. */
     "baseUrl": "./", /* Specify the base directory to resolve non-relative module names. */
     "sourceMap": true, /* Create source map files for emitted JavaScript files. */
-    "outDir": "./dist", /* Specify an output folder for all emitted files. */
+    "outDir": "./dist/esm", /* Specify an output folder for all emitted files. */
     "esModuleInterop": true, /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables 'allowSyntheticDefaultImports' for type compatibility. */
     "forceConsistentCasingInFileNames": true, /* Ensure that casing is correct in imports. */
     "strict": true, /* Enable all strict type-checking options. */
     "skipLibCheck": true /* Skip type checking all .d.ts files. */,
     "resolveJsonModule": true,
+    "declarationMap": true,
+    "inlineSourceMap": false,
+    "lib": [
+      "ES2020",
+      "DOM"
+    ],
     "paths": {
       "@/*": [
         "src/*"
